backend/requirements.txt
asgiref==3.11.1
channels==4.3.2
daphne==4.2.1
dj-database-url==2.2.0
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==26.0
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.7.0
